requirements.txt
Django==4.1.6
psycopg2-binary==2.9.5
django-taggit==3.1.0
pillow==9.4.0